Rename modules and packages with PEP8
