+++ rss-reader/storage/__init__.py
@@ -1,0 +1,1 @@
+from storage.controller import *
