--- rss-reader/storage_controller/__init__.py
@@ -1,1 +1,0 @@
-from storage_controller.controller import *
