--- rss-reader/storage_controller/managers/__init__.py
@@ -1,2 +1,0 @@
-from .articlemanager import ArticleManager, DB_HANDLE
-from .sourcemanager import SourceManager, DB_HANDLE
